+++ src/FloatToIntRepresentation.cpp
@@ -2,25 +2,25 @@
 
 #include <vector>
 
-float UpscaleFloatToIntAndReturnMultiplicator(std::vector<float>& floats, std::vector<epic::longUInt>& ints) {
+double UpscaleFloatToIntAndReturnMultiplicator(std::vector<double>& floats, std::vector<epic::longUInt>& ints) {
 	ints.resize(floats.size());
 
 	bool factorIsFound = false;
 
 	// definitions for rounding accuracy
-	float multiplicator = 1.0f;
+	double multiplicator = 1.0;
 	// how much the multiplicator increases after each check
-	float incrementor = 0.5f;
+	double incrementor = 0.5;
 	// maximum difference for each float from the next full number
-	float maxDifference = 0.1f;
+	double maxDifference = 0.1;
 	// placeholder variable for computing differences
-	float difference = 0.0;
+	double difference = 0.0;
 
 	for (int i = 0; i < 100000 && !factorIsFound; i++) {
 		multiplicator += incrementor;
 		factorIsFound = true;
 		for (auto it : floats) {
-			difference = (it * multiplicator) - (float)((int)(it * multiplicator));
+			difference = (it * multiplicator) - (double)((int)(it * multiplicator));
 			if (difference > maxDifference || difference < -maxDifference) {
 				factorIsFound = false;
 			}
